Move unused prepare_bonded_system to attic
