+++ attic/testsystems/bonded.py
@@ -1,0 +1,48 @@
+import numpy as np
+from timemachine import potentials
+from timemachine.potentials import bonded
+import functools
+
+def prepare_bonded_system(x, B, A, T, precision):
+
+    assert x.ndim == 2
+    N = x.shape[0]
+    # D = x.shape[1]
+
+    atom_idxs = np.arange(N)
+
+    bond_params = np.random.rand(B, 2).astype(np.float64)
+    bond_idxs = []
+    for _ in range(B):
+        bond_idxs.append(np.random.choice(atom_idxs, size=2, replace=False))
+    bond_idxs = np.array(bond_idxs, dtype=np.int32)
+    # params = np.concatenate([params, bond_params])
+
+    # angle_params = np.random.rand(P_angles).astype(np.float64)
+    # angle_param_idxs = np.random.randint(low=0, high=P_angles, size=(A,2), dtype=np.int32) + len(params)
+    # angle_idxs = []
+    # for _ in range(A):
+    #     angle_idxs.append(np.random.choice(atom_idxs, size=3, replace=False))
+    # angle_idxs = np.array(angle_idxs, dtype=np.int32)
+    # params = np.concatenate([params, angle_params])
+
+    # torsion_params = np.random.rand(P_torsions).astype(np.float64)
+    # torsion_param_idxs = np.random.randint(low=0, high=P_torsions, size=(T,3), dtype=np.int32) + len(params)
+    # torsion_idxs = []
+    # for _ in range(T):
+    #     torsion_idxs.append(np.random.choice(atom_idxs, size=4, replace=False))
+    # torsion_idxs = np.array(torsion_idxs, dtype=np.int32)
+    # params = np.concatenate([params, torsion_params])
+
+    print("precision", precision)
+    custom_bonded = potentials.HarmonicBond(bond_idxs, bond_params, precision=precision)
+    harmonic_bond_fn = functools.partial(bonded.harmonic_bond, box=None, bond_idxs=bond_idxs)
+
+    # custom_angles = potentials.HarmonicAngle(angle_idxs, angle_param_idxs, precision=precision)
+    # harmonic_angle_fn = functools.partial(bonded.harmonic_angle, box=None, angle_idxs=angle_idxs, param_idxs=angle_param_idxs)
+
+    # custom_torsions = potentials.PeriodicTorsion(torsion_idxs, torsion_param_idxs, precision=precision)
+    # periodic_torsion_fn = functools.partial(bonded.periodic_torsion, box=None, torsion_idxs=torsion_idxs, param_idxs=torsion_param_idxs)
+
+    return (bond_params, harmonic_bond_fn), custom_bonded
+    # return params, [harmonic_bond_fn, harmonic_angle_fn, periodic_torsion_fn], [custom_bonded, custom_angles, custom_torsions]
